Все работает, покупка и бд реализованы
